undo ingress changes
